requirements.txt
groq
pypdf
python-docx
customtkinter
streamlit